+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Important keybinds now shown at the bottom of each view
+
 ## [0.2.0] - 2025-04-09
 
 ### Added