+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2025-04-13
+
 ### Added
 - Important keybinds now shown at the bottom of each view
 - Add a 'details' view for each anime to see more information about it